refactor tests, and cache decorator
